+++ packages/eslint-config-edb/index.js
@@ -5,7 +5,6 @@ module.exports = {
     "@next/next/no-html-link-for-pages": "off",
     "turbo/no-undeclared-env-vars": "off",
     "simple-import-sort/imports": "error",
-    semi: ["error", "always"],
     "no-console": ["error", { allow: ["warn", "error"] }],
     "no-debugger": "error",
     "no-alert": "error",